test: add useRequest's plugins test case
